# quick-cart-backend/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Quick-Cart Multi-Vendor E-Commerce API
  description: |
    # Quick-Cart API Documentation

    Welcome to the Quick-Cart Multi-Vendor E-Commerce Platform API. This API enables you to build powerful e-commerce applications with multi-vendor support, complete with role-based access control.

    ## 🔐 Authentication

    To test protected endpoints:
    1. Click the **"Authorize"** button (🔓 icon) at the top right
    2. Login using POST `/api/v1/auth/login` to get your JWT token
    3. Enter the token in the format: `Bearer <your_token_here>`
    4. Click **"Authorize"** to apply
    5. Now you can test all protected endpoints!

    ### Test Credentials
    Create an account first, or ask admin for credentials.

    ## 🎭 Roles & Permissions

    | Role | Description | Key Permissions |
    |------|-------------|-----------------|
    | **SUPER_ADMIN** | Platform administrator | All permissions |
    | **VENDOR** | Store owner/seller | Manage own products, view own orders |
    | **CUSTOMER** | Regular user | Browse, purchase, manage own orders |
    | **SUPPORT** | Customer support | View users, orders, products |

    ## 📝 Permission System

    Each endpoint shows required permissions in its description:
    - 🔓 **Public** - No authentication required
    - 🔒 **Protected** - Authentication required
    - 🛡️ **Permission Required** - Specific permission needed
    - 👤 **Ownership** - Must own the resource

  version: 1.0.0
  contact:
...
  license:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Enter your JWT token in the format: `Bearer <token>`

        **How to get a token:**
        1. Register or login using the auth endpoints
        2. Copy the `accessToken` from the response
        3. Click "Authorize" button above
        4. Paste token and click "Authorize"
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "Error description"
      required:
        - success
        - message
...
paths:
...
  /api/v1/auth/register:
    post:
      tags:
        - 🔐 Authentication
      summary: Register new user
      description: |
        🔓 **Public Endpoint** - No authentication required

        Register a new customer or vendor account. Vendors require admin approval before they can start selling.

        ### Business Rules:
        - Email must be unique
        - Password minimum 8 characters
        - Vendors are created with `isApproved: false`
        - Customers are immediately active

      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
                - role
              properties:
                name:
                  type: string
                  example: "John Doe"
                  minLength: 2
                email:
                  type: string
                  format: email
                  example: "john@example.com"
                password:
                  type: string
                  format: password
                  minLength: 8
                  example: "SecurePass123!"
                role:
                  type: string
                  enum: [CUSTOMER, VENDOR]
                  example: "CUSTOMER"
            examples:
              customer:
                summary: Register as Customer
                value:
                  name: "Jane Smith"
                  email: "jane@example.com"
                  password: "MyPassword123"
                  role: "CUSTOMER"
              vendor:
                summary: Register as Vendor
                value:
                  name: "Tech Store Owner"
                  email: "vendor@techstore.com"
                  password: "VendorPass123"
                  role: "VENDOR"
                  businessName: "Tech Store"
                  phoneNumber: "+1234567890"
      responses:
        "201":
          description: ✅ User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Customer account created successfully"
              examples:
                customer:
                  summary: Customer Registration
                  value:
                    message: "Customer account created successfully"
                vendor:
                  summary: Vendor Registration
                  value:
                    message: "Seller account created successfully, pending approval."
        "400":
          description: ❌ Bad Request - Validation error or user exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                userExists:
                  summary: User Already Exists
                  value:
                    success: false
                    message: "User already exists"
                invalidData:
                  summary: Invalid Input
                  value:
                    success: false
                    message: "Validation failed"
        "500":
          description: ❌ Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/v1/auth/login:
    post:
      tags:
        - 🔐 Authentication
      summary: User login
      description: |
        🔓 **Public Endpoint** - No authentication required

        Authenticate with email and password to receive JWT tokens.

        ### Response Tokens:
        - **accessToken**: Use for API requests (expires in 7 days)
        - **refreshToken**: Use to get new access token

        ### How to use the token:
        1. Copy the `accessToken` from response
        2. Click "Authorize" button (🔓 top right)
        3. Enter: `Bearer <your_token>`
        4. Test protected endpoints!

      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "john@example.com"
                password:
                  type: string
                  format: password
                  example: "SecurePass123!"
      responses:
        "200":
          description: ✅ Login successful
          headers:
            Set-Cookie:
              description: HTTP-only cookie with refresh token
              schema:
                type: string
                example: "token=eyJhbGc...; HttpOnly; Max-Age=604800"
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  email:
                    type: string
                  role:
                    type: string
                  avatar:
                    type: string
                    nullable: true
              example:
                id: "507f1f77bcf86cd799439011"
                name: "John Doe"
                email: "john@example.com"
                role: "CUSTOMER"
                avatar: null
        "401":
          description: ❌ Unauthorized - Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                message: "Invalid Credentials!"
        "500":
          description: ❌ Server Error
  /api/v1/auth/logout:
    post:
      tags:
        - 🔐 Authentication
      summary: User logout
      description: |
        🔒 **Protected Endpoint** - Authentication required

        Clears the authentication cookie and logs out the user.
      operationId: logout
      security:
        - BearerAuth: []
      responses:
        "200":
          description: ✅ Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logout Successful"
  /api/v1/auth/login/admin:
    post:
      tags:
        - 🔐 Authentication
      summary: Admin/Super Admin login
      description: |
        🔓 **Public Endpoint** - No authentication required

        Special login endpoint for admin and super admin users only.
        Regular users and vendors cannot use this endpoint.

        **Allowed Roles:**
        - SUPER_ADMIN
        - ADMIN
      operationId: loginAdmin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  example: "admin@quickcart.com"
                password:
                  type: string
                  example: "AdminPass123!"
      responses:
        "200":
          description: ✅ Admin login successful
        "401":
          description: ❌ Unauthorized - Not an admin or invalid credentials

  # ==================== USERS ====================
...